+++ docs/sections/changelog.rst
@@ -12,6 +12,7 @@ Unreleased
 * Added: request in password grant authenticate call.
 * Changed: dropping support for Django versions before 1.8.
 * Changed: pass token and request to OIDC_IDTOKEN_PROCESSING_HOOK.
+* Fixed: CORS OPTIONS request blocked on userinfo request.
 
 0.6.0
 =====